+++ src/common/constants/token.ts
@@ -2,7 +2,8 @@ import { MAINNET_FOLKS_TOKEN_ID, TESTNET_FOLKS_TOKEN_ID } from "../types/token.j
 
 import type { FolksTokenId } from "../types/token.js";
 
-export const CIRCLE_FOLKS_TOKEN_ID: Array<FolksTokenId> = [MAINNET_FOLKS_TOKEN_ID.USDC, TESTNET_FOLKS_TOKEN_ID.USDC];
-export const CCIP_FOLKS_TOKEN_ID: Array<FolksTokenId> = [TESTNET_FOLKS_TOKEN_ID.CCIP_BnM];
-
-export const BRIDGED_FOLKS_TOKEN_ID: Array<FolksTokenId> = [...CIRCLE_FOLKS_TOKEN_ID, ...CCIP_FOLKS_TOKEN_ID];
+export const CROSS_CHAIN_FOLKS_TOKEN_ID: Array<FolksTokenId> = [
+  MAINNET_FOLKS_TOKEN_ID.USDC,
+  MAINNET_FOLKS_TOKEN_ID.SolvBTC,
+  TESTNET_FOLKS_TOKEN_ID.USDC,
+];