+++ src/common/types/token.ts
@@ -1,4 +1,5 @@
 import type { GenericAddress } from "./address.js";
+import type { AdapterType } from "./message.js";
 
 export const MAINNET_FOLKS_TOKEN_ID = {
   USDC: "USDC",
@@ -13,6 +14,9 @@ export const MAINNET_FOLKS_TOKEN_ID = {
   BNB: "BNB",
   ETHB_bsc: "ETHB_bsc",
   BTCB_bsc: "BTCB_bsc",
+  ETH_arb: "ETH_arb",
+  ARB: "ARB",
+  SolvBTC: "SolvBTC",
 } as const;
 export type MainnetFolksTokenId = (typeof MAINNET_FOLKS_TOKEN_ID)[keyof typeof MAINNET_FOLKS_TOKEN_ID];
 
@@ -33,7 +37,7 @@ export type FolksTokenId = MainnetFolksTokenId | TestnetFolksTokenId;
 export enum TokenType {
   NATIVE = "NATIVE",
   ERC20 = "ERC20",
-  CIRCLE = "CIRCLE",
+  CROSS_CHAIN = "CROSS_CHAIN",
 }
 
 export type ITokenData = {
@@ -51,8 +55,9 @@ type Erc20HubTokenType = {
   decimals: number;
 };
 
-type CircleTokenType = {
-  type: TokenType.CIRCLE;
+export type CrossChainTokenType = {
+  type: TokenType.CROSS_CHAIN;
+  adapters: Array<AdapterType>;
   address: GenericAddress;
   decimals: number;
 };
@@ -62,8 +67,8 @@ type NativeTokenType = {
   decimals: number;
 };
 
-export type FolksSpokeTokenType = Erc20SpokeTokenType | CircleTokenType | NativeTokenType;
-export type FolksHubTokenType = Erc20HubTokenType | CircleTokenType | NativeTokenType;
+export type FolksSpokeTokenType = Erc20SpokeTokenType | CrossChainTokenType | NativeTokenType;
+export type FolksHubTokenType = Erc20HubTokenType | CrossChainTokenType | NativeTokenType;
 
 export type SpokeTokenData = {
   poolId: number;